#39 Add Sync tab and Sync code
Add Sync tab and cool icon.
All synchronization will being done with the Parse API.
As of this commit, users can create/login to synchronization accounts.
I've added the API keys in a constants file that is being gitignored.
The implementation for synchronization is a big step that I will take next commit.