#51 Share button
Had to do some moving around to fit the new share button.
The size of the navigation bar text "Delightful" had to be smaller.
Also, I added a space before and after delightful to center it better
and ensure it isn't cut off on the ends. Serialize data for sharing.